# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car API",
        "description": "API для управления автомобилями, комплектациями, опциями и ценами",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/options": {
            "get": {
                "tags": [
                    "Options"
                ],
                "summary": "Получить список всех опций",
                "operationId": "aba522e39edcbd0a845ebb6cf802a499",
                "responses": {
                    "200": {
                        "description": "Список опций",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Leather Seats"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Options"
                ],
                "summary": "Создать новую опцию",
                "operationId": "321d8300ff36fc5056c2f6f0132602cc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Climate Control"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Опция создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Climate Control"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/options/{id}": {
            "get": {
                "tags": [
                    "Options"
                ],
                "summary": "Получить одну опцию",
                "operationId": "98465114297e0e658b4250d937997608",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID опции",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Опция найдена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Sunroof"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Опция не найдена"
                    }
                }
            },
            "put": {
                "tags": [
                    "Options"
                ],
                "summary": "Обновить опцию",
                "operationId": "9c85be0f9234021de4a019f59b606c71",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID опции",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Option"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Опция обновлена"
                    },
                    "404": {
                        "description": "Опция не найдена"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Options"
                ],
                "summary": "Удалить опцию",
                "operationId": "75240c34e59673b44a9e49f58fe805f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID опции",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Опция удалена"
                    },
                    "404": {
                        "description": "Опция не найдена"
                    }
                }
            }
        },
...
    },
...
}